backend/requirements.txt
assemblyai
cohere
crewai
crewai_tools
duckduckgo-search
faiss-cpu
google-generativeai
jobspy
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-openai
langgraph
langsmith
moviepy
numpy
pandas
pymongo
pypdf
requests
sentence-transformers
torch
tqdm
unstructured
unstructured[pdf]
wikipedia